# src/app/api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: OpenBMC Learning Platform API
  description: |
    REST API for the OpenBMC Learning Platform. Provides endpoints for authentication,
    learning paths, lessons, progress tracking, AI features, and admin operations.
  version: 1.0.0
  contact:
    name: OpenBMC Learning Platform
  license:
    name: MIT
servers:
  - url: /api
    description: API base path
...
paths:
...
  /admin/users:
    get:
      tags: [Admin]
      summary: List all users
      operationId: listUsers
      security:
        - cookieAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: role
          in: query
          schema:
            type: string
            enum: [LEARNER, EDITOR, ADMIN]
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      users:
                        type: array
                        items:
                          $ref: '#/components/schemas/AdminUser'
                      total:
                        type: integer
                      page:
                        type: integer
                      pageSize:
                        type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Admin role required
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: JWT authentication token stored in HTTP-only cookie
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
...
    AdminUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        displayName:
          type: string
          nullable: true
        role:
          type: string
        createdAt:
          type: string
          format: date-time
        emailVerified:
          type: string
          format: date-time
          nullable: true
...